Working Build
Fixed control logic for how pwmLED value was updated, behavior is now smoother and PID correction occurs even if no user inputs are detected. Center button prints out the current PID configuration, fixed underflow issue with gain and setpoint, added 2 tick delay to slowdown how often the serial port is written; Serial port is still written often and impossible to see the output from the PID when the center button is pressed, would need a longer delay or to put all of the print functions under user control. Needed at least a tick delay to prevent the PIDtask from over burdening the CPU resources and to allow the interrupts to be processed

+++ src/pidtask.h
@@ -33,7 +33,7 @@ bool pid_init (PID_t* pid);
 // prototype of pid function that returns the controlled signal as a float
 // takes lux_value, returned from TSL2561 sensor returns float vlaue 
 // is used to determine what value to write for LED PWM 
-float pid_funct (PID_t* pid, float lux_value, uint8_t switches);
+float pid_funct (PID_t* pid, uint16_t lux_value, uint8_t switches);
 
 /*********************PID Task Prototype*************************************
 *   Task Handles the Following:
